# swagger.json
{
  "swagger": "3.0.0",
  "openapi": "3.0.0",
  "info": {
    "title": "Swagger documentstion - OpenAPI 3.0",
    "description": "My brand api documention",
    "termsOfService": "https://nsanzimfura.up.railway.app/",
    "contact": {
      "email": "erickykress1gmail.com",
      "Tel": "0780088171"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "https://nsanzimfura-server.up.railway.app/"
    },
    {
      "url": "http://localhost:5000/"
    }
  ],
  "schemes": ["https", "http"],
  "paths": {
...
    "/mybrand/api/users/{id}": {
      "get": {
        "tags": ["user"],
        "summary": "Get user by Id",
        "description": "",
        "operationId": "getUserByName",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The name that needs to be fetched. Use user1 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["user"],
        "summary": "Update user",
        "description": "This can only be done by the logged in Admin.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The name that needs to be fetched. Use user1 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update an existent user in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          },
          "404": {
            "description": "The user with this Id not found!"
          }
        }
      },
      "delete": {
        "tags": ["user"],
        "summary": "Delete user",
        "description": "User is deleted from the system.",
        "operationId": "deleteUser",
        "parameters": [
          {
            "title": "id",
            "in": "path",
            "required": true,
            "description": "ID of blog that we want to find",
            "type": "string"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "202": "Deleted successfully",
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/mybrand/api/blogs/{id}": {
      "parameters": [
        {
          "title": "id",
          "in": "path",
          "required": true,
          "description": "ID of blog that we want to find",
          "type": "string"
        }
      ],
      "delete": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete blog with given ID",
        "tags": ["blogs"],
        "responses": {
          "200": {
            "description": "Blog is deleted",
            "schema": {}
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": ["blogs"],
        "summary": "Update Blog",
        "description": "This can only be done by the logged in Admin.",
        "operationId": "updateUser",
        "requestBody": {
          "description": "Update an existent blog in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/blog"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "successful operation"
          },
          "404": {
            "description": "The user with this Id not found!"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "nsanzimfura eric"
          },
          "email": {
            "type": "string",
            "example": "erickres3s@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "kress123"
          },
          "confirmPass": {
            "type": "string",
            "example": "kress123"
          }
        },
        "xml": {
          "name": "user"
        }
      },
...
      "blog": {
        "required": ["title", "description", "img", "category"],
        "properties": {
          "title": {
            "type": "string",
            "example":"string"
          },
          "description": {
            "type": "string",
            "example":"Lorem ipsum dolor sit ameet"
          },
          "category": {
            "type": "string",
            "example":"Code"
          },
          "img": {
            "type": "string",
            "example":"img url"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "security": {
        "bearerAuth": []
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}